Restructure epics
divide epics according to domain
